Create getStream.py
Here is my attempt to get real time streaming data from twitter using tweepy, The result shows all the tweets hashtagged "#Halloween", and it write the result a json file in text. We can also decode them and filter some useless characters directly, but we  can do that in data processing and analysis.